Coerce initializer parameter to_hash or to_h when it's not a Hash
Passing hash-like objects that respond to #to_h or #to_hash will no longer fail to
work with lift's initializer when they didn't respond to #each_pair. Whenever we can
coerce the object to be a hash, do it, rather than forcing the caller to.

This can be used in situations like passing a JSON API response directly into the
initializer of a class that includes Lift, to create a value object out of the JSON
data.